backend/requirements.txt
asgiref==3.5.0
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
defusedxml==0.7.1
Django==3.2
django-cors-headers==3.11.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==2.10
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
oauthlib==3.2.0
olefile==0.46
Pillow==8.4.0
pip==21.2.4
pipfile==0.0.2
pycparser==2.20
PyJWT==2.3.0
pyOpenSSL==19.1.0
PySocks==1.7.1
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
setuptools==58.0.4
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
stripe==2.49.0
toml==0.10.2
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.25.11
wheel==0.37.1
